file 레파지토리 테스트 완료

+++ Desktop/sporting/sporting_server/src/main/java/shop/mtcoding/sporting_server/modules/sport_category/repository/SportCategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import shop.mtcoding.sporting_server.modules.sport_category.entity.SportCategory;
 
-public interface SportCategoryRepository extends JpaRepository<SportCategory, Integer>, SportCategoryCustomRepository {
+public interface SportCategoryRepository extends JpaRepository<SportCategory, Long>, SportCategoryCustomRepository {
 
 }
